Add confirm and instruct prompts, avoid use of futures
This patch implements a blocking response queue for certain messages sent to
the server.  These messages are confirmPrompt, prompt, and instructPrompt.

This makes tornado's yield-based futures redundant, simplifying the tests quite
a bit and makes it possible to follow the same patterns for failing tests in
the harness code as MinimalTestCase uses.